fixup! ARC: Emulate one-byte and two-byte cmpxchg

+++ arch/arc/include/asm/cmpxchg.h
@@ -50,7 +50,7 @@
 		_prev_ = cmpxchg_emu_u8((volatile u8 *)_p_, _o_, _n_);	\
 		break;							\
 	case 2:								\
-		_prev_ = cmpxchg_emu_u16((volatile u8 *)_p_, _o_, _n_);	\
+		_prev_ = cmpxchg_emu_u16((volatile u16 *)_p_, _o_, _n_); \
 		break;							\
 	case 4:								\
 		_prev_ = __cmpxchg(_p_, _o_, _n_);			\
@@ -77,7 +77,7 @@
 		_prev_ = (__typeof__(*(ptr)))__flags;			\
 		break;							\
 	case 2:								\
-		__flags = cmpxchg_emu_u16((volatile u8 *)_p_, _o_, _n_); \
+		__flags = cmpxchg_emu_u16((volatile u16 *)_p_, _o_, _n_); \
 		_prev_ = (__typeof__(*(ptr)))__flags;			\
 		break;							\
 	case 4:								\
